csdl2/requirements.txt
prettytable
mysql-connector-python
customtkinter
